requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.4.0
attrs==21.2.0
certifi==2020.6.20
chardet==3.0.4
click==8.0.1
colorama==0.4.4
Flask==1.0.2
idna==2.10
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==2.10
lml==0.1.0
MarkupSafe==2.0.1
more-itertools==8.8.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
pyparsing==2.4.7
pytest==4.5.0
pytest-html==1.19.0
pytest-metadata==1.11.0
PyTestReport==0.2.1
PyYAML==5.4.1
requests==2.24.0
six==1.16.0
toml==0.10.2
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==2.0.1
xlrd==1.2.0
xlwt==1.3.0
